+++ src/thin_client.rs
@@ -297,8 +297,8 @@ mod tests {
     #[test]
     fn test_thin_client() {
         logger::setup();
-        let leader_kp = KeyPair::new();
-        let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+        let leader_keypair = KeyPair::new();
+        let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
         let leader_data = leader.data.clone();
 
         let alice = Mint::new(10_000);
@@ -307,7 +307,7 @@ mod tests {
         let exit = Arc::new(AtomicBool::new(false));
 
         let server = FullNode::new_leader(
-            leader_kp,
+            leader_keypair,
             bank,
             0,
             None,
@@ -343,8 +343,8 @@ mod tests {
     #[ignore]
     fn test_bad_sig() {
         logger::setup();
-        let leader_kp = KeyPair::new();
-        let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+        let leader_keypair = KeyPair::new();
+        let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
         let alice = Mint::new(10_000);
         let bank = Bank::new(&alice);
         let bob_pubkey = KeyPair::new().pubkey();
@@ -352,7 +352,7 @@ mod tests {
         let leader_data = leader.data.clone();
 
         let server = FullNode::new_leader(
-            leader_kp,
+            leader_keypair,
             bank,
             0,
             None,
@@ -400,15 +400,15 @@ mod tests {
     #[test]
     fn test_client_check_signature() {
         logger::setup();
-        let leader_kp = KeyPair::new();
-        let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+        let leader_keypair = KeyPair::new();
+        let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
         let alice = Mint::new(10_000);
         let bank = Bank::new(&alice);
         let bob_pubkey = KeyPair::new().pubkey();
         let exit = Arc::new(AtomicBool::new(false));
         let leader_data = leader.data.clone();
         let server = FullNode::new_leader(
-            leader_kp,
+            leader_keypair,
             bank,
             0,
             None,
