+++ tests/multinode.rs
@@ -91,9 +91,9 @@ fn test_multi_node_validator_catchup_from_zero() {
     logger::setup();
     const N: usize = 5;
     trace!("test_multi_node_validator_catchup_from_zero");
-    let leader_kp = KeyPair::new();
-    let leader_pubkey = leader_kp.pubkey().clone();
-    let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+    let leader_keypair = KeyPair::new();
+    let leader_pubkey = leader_keypair.pubkey().clone();
+    let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
     let leader_data = leader.data.clone();
     let bob_pubkey = KeyPair::new().pubkey();
 
@@ -102,7 +102,7 @@ fn test_multi_node_validator_catchup_from_zero() {
         leader,
         true,
         LedgerFile::Path(ledger_path.clone()),
-        leader_kp,
+        leader_keypair,
         None,
     );
 
@@ -199,9 +199,9 @@ fn test_multi_node_basic() {
     logger::setup();
     const N: usize = 5;
     trace!("test_multi_node_basic");
-    let leader_kp = KeyPair::new();
-    let leader_pubkey = leader_kp.pubkey().clone();
-    let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+    let leader_keypair = KeyPair::new();
+    let leader_pubkey = leader_keypair.pubkey().clone();
+    let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
     let leader_data = leader.data.clone();
     let bob_pubkey = KeyPair::new().pubkey();
     let (alice, ledger_path) = genesis(10_000);
@@ -209,7 +209,7 @@ fn test_multi_node_basic() {
         leader,
         true,
         LedgerFile::Path(ledger_path.clone()),
-        leader_kp,
+        leader_keypair,
         None,
     );
 
@@ -260,8 +260,8 @@ fn test_multi_node_basic() {
 #[test]
 fn test_boot_validator_from_file() {
     logger::setup();
-    let leader_kp = KeyPair::new();
-    let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+    let leader_keypair = KeyPair::new();
+    let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
     let bob_pubkey = KeyPair::new().pubkey();
     let (alice, ledger_path) = genesis(100_000);
     let leader_data = leader.data.clone();
@@ -269,7 +269,7 @@ fn test_boot_validator_from_file() {
         leader,
         true,
         LedgerFile::Path(ledger_path.clone()),
-        leader_kp,
+        leader_keypair,
         None,
     );
     let leader_balance =
@@ -299,14 +299,14 @@ fn test_boot_validator_from_file() {
 }
 
 fn create_leader(ledger_path: &str) -> (NodeInfo, FullNode) {
-    let leader_kp = KeyPair::new();
-    let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+    let leader_keypair = KeyPair::new();
+    let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
     let leader_data = leader.data.clone();
     let leader_fullnode = FullNode::new(
         leader,
         true,
         LedgerFile::Path(ledger_path.to_string()),
-        leader_kp,
+        leader_keypair,
         None,
     );
     (leader_data, leader_fullnode)
@@ -405,9 +405,9 @@ fn test_multi_node_dynamic_network() {
         Err(_) => std::usize::MAX,
     };
 
-    let leader_kp = KeyPair::new();
-    let leader_pubkey = leader_kp.pubkey().clone();
-    let leader = TestNode::new_localhost_with_pubkey(leader_kp.pubkey());
+    let leader_keypair = KeyPair::new();
+    let leader_pubkey = leader_keypair.pubkey().clone();
+    let leader = TestNode::new_localhost_with_pubkey(leader_keypair.pubkey());
     let bob_pubkey = KeyPair::new().pubkey();
     let (alice, ledger_path) = genesis(10_000_000);
     let alice_arc = Arc::new(RwLock::new(alice));
@@ -416,7 +416,7 @@ fn test_multi_node_dynamic_network() {
         leader,
         true,
         LedgerFile::Path(ledger_path.clone()),
-        leader_kp,
+        leader_keypair,
         None,
     );
